Initial move
Move all worksheet's stuff from bika/lims/browser to bik/lims/browser/worksheet
