+++ bika/lims/browser/worksheet/templates/add_analyses.pt
