+++ bika/lims/browser/worksheet/worksheets.py
@@ -1,3 +1,4 @@
+# coding=utf-8
 from DateTime import DateTime
 from Products.Archetypes.config import REFERENCE_CATALOG
 from Products.Archetypes.public import DisplayList
@@ -21,56 +22,6 @@ from zope.interface import implements
 import plone
 import json
 import zope
-
-class WorksheetFolderWorkflowAction(WorkflowAction):
-    """ Workflow actions taken in the WorksheetFolder
-        This function is called to do the workflow actions
-        that apply to worksheets in the WorksheetFolder
-    """
-    def __call__(self):
-        form = self.request.form
-        plone.protect.CheckAuthenticator(form)
-        workflow = getToolByName(self.context, 'portal_workflow')
-        rc = getToolByName(self.context, REFERENCE_CATALOG)
-        action, came_from = WorkflowAction._get_form_workflow_action(self)
-
-        if action == 'reassign':
-            mtool = getToolByName(self.context, 'portal_membership')
-            if not mtool.checkPermission(ManageWorksheets, self.context):
-
-                # Redirect to WS list
-                msg = _('You do not have sufficient privileges to '
-                        'manage worksheets.')
-                self.context.plone_utils.addPortalMessage(msg, 'warning')
-                portal = getToolByName(self.context, 'portal_url').getPortalObject()
-                self.destination_url = portal.absolute_url() + "/worksheets"
-                self.request.response.redirect(self.destination_url)
-                return
-
-            selected_worksheets = WorkflowAction._get_selected_items(self)
-            selected_worksheet_uids = selected_worksheets.keys()
-
-            if selected_worksheets:
-                changes = False
-                for uid in selected_worksheet_uids:
-                    worksheet = selected_worksheets[uid]
-                    # Double-check the state first
-                    if workflow.getInfoFor(worksheet, 'review_state') == 'open':
-                        worksheet.setAnalyst(form['Analyst'][0][uid])
-                        worksheet.reindexObject(idxs=['getAnalyst'])
-                        changes = True
-
-                if changes:
-                    message = PMF('Changes saved.')
-                    self.context.plone_utils.addPortalMessage(message, 'info')
-
-            self.destination_url = self.request.get_header("referer",
-                                   self.context.absolute_url())
-            self.request.response.redirect(self.destination_url)
-        else:
-            # default bika_listing.py/WorkflowAction for other transitions
-            WorkflowAction.__call__(self)
-
 
 class WorksheetFolderListingView(BikaListingView):
 
@@ -558,3 +509,53 @@ class AddWorksheetView(BrowserView):
             msg = _("No analyses were added")
             self.context.plone_utils.addPortalMessage(msg)
             self.request.RESPONSE.redirect(ws.absolute_url() + "/add_analyses")
+
+
+class WorksheetFolderWorkflowAction(WorkflowAction):
+    """ Workflow actions taken in the WorksheetFolder
+        This function is called to do the workflow actions
+        that apply to worksheets in the WorksheetFolder
+    """
+    def __call__(self):
+        form = self.request.form
+        plone.protect.CheckAuthenticator(form)
+        workflow = getToolByName(self.context, 'portal_workflow')
+        rc = getToolByName(self.context, REFERENCE_CATALOG)
+        action, came_from = WorkflowAction._get_form_workflow_action(self)
+
+        if action == 'reassign':
+            mtool = getToolByName(self.context, 'portal_membership')
+            if not mtool.checkPermission(ManageWorksheets, self.context):
+
+                # Redirect to WS list
+                msg = _('You do not have sufficient privileges to '
+                        'manage worksheets.')
+                self.context.plone_utils.addPortalMessage(msg, 'warning')
+                portal = getToolByName(self.context, 'portal_url').getPortalObject()
+                self.destination_url = portal.absolute_url() + "/worksheets"
+                self.request.response.redirect(self.destination_url)
+                return
+
+            selected_worksheets = WorkflowAction._get_selected_items(self)
+            selected_worksheet_uids = selected_worksheets.keys()
+
+            if selected_worksheets:
+                changes = False
+                for uid in selected_worksheet_uids:
+                    worksheet = selected_worksheets[uid]
+                    # Double-check the state first
+                    if workflow.getInfoFor(worksheet, 'review_state') == 'open':
+                        worksheet.setAnalyst(form['Analyst'][0][uid])
+                        worksheet.reindexObject(idxs=['getAnalyst'])
+                        changes = True
+
+                if changes:
+                    message = PMF('Changes saved.')
+                    self.context.plone_utils.addPortalMessage(message, 'info')
+
+            self.destination_url = self.request.get_header("referer",
+                                   self.context.absolute_url())
+            self.request.response.redirect(self.destination_url)
+        else:
+            # default bika_listing.py/WorkflowAction for other transitions
+            WorkflowAction.__call__(self)
