--- bika/lims/browser/worksheetfolder.zcml
@@ -1,32 +1,0 @@
-<configure
-    xmlns="http://namespaces.zope.org/zope"
-    xmlns:browser="http://namespaces.zope.org/browser"
-    xmlns:i18n="http://namespaces.zope.org/i18n"
-    i18n_domain="bika">
-
-    <browser:page
-      for="bika.lims.interfaces.IWorksheetFolder"
-      name="base_view"
-      class="bika.lims.browser.worksheetfolder.WorksheetFolderListingView"
-      permission="zope.Public"
-      layer="bika.lims.interfaces.IBikaLIMS"
-    />
-
-    <browser:page
-      for="bika.lims.interfaces.IWorksheetFolder"
-      name="worksheet_add"
-      class="bika.lims.browser.worksheetfolder.AddWorksheetView"
-      permission="zope.Public"
-      layer="bika.lims.interfaces.IBikaLIMS"
-    />
-
-    <!-- Workflow action button clicked -->
-    <browser:page
-      for="bika.lims.interfaces.IWorksheetFolder"
-      name="workflow_action"
-      class="bika.lims.browser.worksheetfolder.WorksheetFolderWorkflowAction"
-      permission="zope.Public"
-      layer="bika.lims.interfaces.IBikaLIMS"
-    />
-
-</configure>
